V0.0.1 Completada

+++ src/app/components/tasks/tasks.component.ts
@@ -3,6 +3,7 @@ import { Component, OnInit } from '@angular/core';
 /* */
 import { Task } from '../../models/task';
 import { TaskService } from '../../services/task.service';
+import { log } from 'util';
 
 @Component({
   selector: 'app-tasks',
@@ -13,6 +14,10 @@ export class TasksComponent implements OnInit {
   /* */
   tasks: Task[];
   title: string;
+
+  editState: Boolean = false;
+  taskToEdit: Task;
+
   constructor(/* */ public taskService: TaskService) { }
 
   /*Primer metodo que se ejecuta*/
@@ -23,8 +28,22 @@ export class TasksComponent implements OnInit {
     });
   }
 
-  deleteTask($event, task) {
-    this.taskService.deleteTask(task);
+  deleteTask(event, task) {
+    const response = confirm('are you sure you want delete?');
+    if (response === true) {
+      this.taskService.deleteTask(task);
+    }
+    return;
   }
 
+  editTask(event, task) {
+    this.editState = !this.editState;
+    this.taskToEdit = task;
+  }
+
+  updateTask(task) {
+    this.taskService.updateTask(task);
+    this.taskToEdit = null;
+    this.editState = false;
+  }
 }
